breakout PathfindingSwapButton
