+++ cmd/ui/src/views/Explore/ExploreSearch/PathfindingSwapButton.tsx
@@ -1,0 +1,66 @@
+// Copyright 2023 Specter Ops, Inc.
+//
+// Licensed under the Apache License, Version 2.0
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+import { faExchangeAlt } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { Button } from '@mui/material';
+import makeStyles from '@mui/styles/makeStyles';
+import { useCallback } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { setSearchValue, startSearchSelected } from 'src/ducks/searchbar/actions';
+import { PATHFINDING_SEARCH, PRIMARY_SEARCH, SEARCH_TYPE_EXACT } from 'src/ducks/searchbar/types';
+import { AppState } from 'src/store';
+
+const useStyles = makeStyles((theme) => ({
+    swapButton: {
+        height: '25px',
+        width: '25px',
+        minWidth: '25px',
+        borderRadius: theme.shape.borderRadius,
+        borderColor: 'rgba(0,0,0,0.23)',
+        color: 'black',
+        padding: 0,
+    },
+}));
+
+const PathfindingSwapButton = () => {
+    const classes = useStyles();
+    const dispatch = useDispatch();
+
+    const { primary, secondary } = useSelector((state: AppState) => state.search);
+
+    const swapPathfindingInputs = useCallback(() => {
+        const newSourceNode = secondary.value;
+        const newDestinationNode = primary.value;
+
+        dispatch(setSearchValue(newSourceNode, PRIMARY_SEARCH, SEARCH_TYPE_EXACT));
+        dispatch(setSearchValue(newDestinationNode, PATHFINDING_SEARCH, SEARCH_TYPE_EXACT));
+
+        dispatch(startSearchSelected(PATHFINDING_SEARCH));
+    }, [primary, secondary, dispatch]);
+
+    return (
+        <Button
+            className={classes.swapButton}
+            variant='outlined'
+            disabled={!primary?.value || !secondary?.value}
+            onClick={swapPathfindingInputs}>
+            <FontAwesomeIcon icon={faExchangeAlt} className='fa-rotate-90' />
+        </Button>
+    );
+};
+
+export default PathfindingSwapButton;
